PfaffianSystems

Documentation for PfaffianSystems.

AbstractAlgebra.vars

PfaffianSystems.diff_op_ring

PfaffianSystems.dvars

PfaffianSystems.isAsirAvailable

PfaffianSystems.leading_term

PfaffianSystems.normalform

PfaffianSystems.parseAsir

PfaffianSystems.pfaffian_system

PfaffianSystems.runAsir

PfaffianSystems.stdmon

PfaffianSystems.vec2str

PfaffianSystems.weyl_algebra

PfaffianSystems.wnormalform

AbstractAlgebra.vars — Method

vars(wae::T) where T <: AbstractDiffOp

Examples

julia> D, (x, y), (dx, dy) = weyl_algebra(["x", "y"])
julia> vars(x+y)
2-element Vector{PfaffianSystems.WAlgElem{AbstractAlgebra.Generic.MPoly{AbstractAlgebra.Generic.MPoly{Rational{BigInt}}}}}:
 x
 y

julia> R, (x, y), (dx, dy) = diff_op_ring(["x", "y"])
julia> vars(x+y)
2-element Vector{PfaffianSystems.DORElem{AbstractAlgebra.Generic.MPoly{AbstractAlgebra.Generic.RationalFunctionFieldElem{Rational{BigInt}, AbstractAlgebra.Generic.MPoly{Rational{BigInt}}}}}}:
 x
 y

PfaffianSystems.diff_op_ring — Method

Ring of differential operators over rational functions

PfaffianSystems.dvars — Method

dvars(wae::T) where T <: AbstractDiffOp

Examples

julia> D, (x, y), (dx, dy) = weyl_algebra(["x", "y"])
julia> dvars(dx+dy)
2-element Vector{PfaffianSystems.WAlgElem{AbstractAlgebra.Generic.MPoly{AbstractAlgebra.Generic.MPoly{Rational{BigInt}}}}}:
 dx
 dy

julia> R, (x, y), (dx, dy) = diff_op_ring(["x", "y"])
julia> dvars(dx+dy)
2-element Vector{PfaffianSystems.DORElem{AbstractAlgebra.Generic.MPoly{AbstractAlgebra.Generic.RationalFunctionFieldElem{Rational{BigInt}, AbstractAlgebra.Generic.MPoly{Rational{BigInt}}}}}}:
 dx
 dy

PfaffianSystems.isAsirAvailable — Method

Return the version of available Asir as an integer if exists, and otherwise return nothing.

PfaffianSystems.leading_term — Function

leading_term(f::DORElem, order::Symbol=:lex)

Return the leading term of f with respect to order. Only order=:lex is supported now.

PfaffianSystems.normalform — Method

normalform(f::T, G::Vector{T}) where T<: DORElem

Compute normal form of f with respect to G and return the remainder r and quotients q.

PfaffianSystems.parseAsir — Method

parseAsir(asir_res::String)

Add square brackets and separete the response into lines.

PfaffianSystems.pfaffian_system — Method

pfaffian_system(G::Vector{T}, S::Vector{T}) where T <: DORElem

Examples

julia> R, (x, y), (dx, dy) = diff_op_ring(["x", "y"])
julia> pfaffian_system([dx^2 + 1, dy^2 + 1], [one(dx), dx, dy])
2-element Vector{Matrix{AbstractAlgebra.Generic.RationalFunctionFieldElem{Rational{BigInt}, AbstractAlgebra.Generic.MPoly{Rational{BigInt}}}}}:
 [0 1 0; -1 0 0; 0 0 0]
 [0 0 1; 0 0 0; -1 0 0]

PfaffianSystems.runAsir — Method

runAsir(commands::AbstractString)

Run commands on Asir. The raw response of Asir is returned as a string.

PfaffianSystems.stdmon — Method

stdmon!(I::DIdeal[, ordered_vars::Vector{Num}])

Compute standard monomials if I is zero-dimensional as an ideal of R-ring. The order of varaibles in the computation of Grobner bases can be provided as ordered_vars, which can be omitted as long as the computational time is not crutial.

PfaffianSystems.vec2str — Method

vec2str(v::AbstractVector; delim=",")
vec2str(v::AbstractDiffOp; delim=",")
vec2str(vs...; delim=",")

Return a string consiting of all elements of v with delimiter delim.

PfaffianSystems.weyl_algebra — Method

WeylAlgebra

PfaffianSystems.wnormalform — Method

wnormalform(f::T, G::Vector{T}) where T<: DORElem Compute weak normal form of f with respect to G and return the remainder r and quotients q.