PfaffianSystems

Documentation for PfaffianSystems.

AbstractAlgebra.coefficients

AbstractAlgebra.evaluate

AbstractAlgebra.evaluate

AbstractAlgebra.exponent_vectors

AbstractAlgebra.monomials

AbstractAlgebra.vars

PfaffianSystems.diff_op_ring

PfaffianSystems.dvars

PfaffianSystems.integration_DIdeal

PfaffianSystems.intersection_DIdeals

PfaffianSystems.isAsirAvailable

PfaffianSystems.leading_term

PfaffianSystems.multiplication_DIdeal

PfaffianSystems.normalform

PfaffianSystems.parseAsir

PfaffianSystems.pfaffian_system

PfaffianSystems.restriction_DIdeal

PfaffianSystems.runAsir

PfaffianSystems.stdmon

PfaffianSystems.vec2str

PfaffianSystems.weyl_algebra

PfaffianSystems.wnormalform

AbstractAlgebra.coefficients — Method

coefficients(f::T) where T <: AbstractDiffOp

Extract the coefficients of a differential operator.

Arguments:

f::T: A differential operator to extract the coefficients from.

Returns:

An array of instances of T representing the coefficients of the differential operator.

AbstractAlgebra.evaluate — Method

evaluate(I::DIdeal{T}, vrs::Vector{T}, vals::Vector{T}) where T <: WAlgElem

Return the ideal that is obtained by evaluating I at vrs equal to vals.

#Examples

julia> D, (x, y), (dx, dy) = weyl_algebra(["x", "y"])
(2-d Weyl algebra in [x,y], PfaffianSystems.WAlgElem{AbstractAlgebra.Generic.MPoly{AbstractAlgebra.Generic.MPoly{Rational{BigInt}}}}[x, y], PfaffianSystems.WAlgElem{AbstractAlgebra.Generic.MPoly{AbstractAlgebra.Generic.MPoly{Rational{BigInt}}}}[dx, dy])
julia> I = DIdeal(D, [dx + x])
Ideal of 2-d Weyl algebra in [x,y] generated by [dx + x]
julia> evaluate(I, [x, dx], [y, dy])
Ideal of 2-d Weyl algebra in [x,y] generated by [dy + y]

AbstractAlgebra.evaluate — Method

evaluate(dop::T, vars::Vector{T}, vals::Vector{T}) where T <: AbstractDiffOp

Evaluate a differential operator with respect to a set of variables and their corresponding values.

Arguments

dop::T: A differential operator to evaluate.

vars::Vector{T}: A vector of variables to evaluate the differential operator with respect to.

vals::Vector{T}: A vector of values corresponding to the variables.

Returns

An instance of T representing the result of evaluating the differential operator.

The vars and vals vectors must have the same length, and each element in vars must be of the same type as its corresponding element in vals.

Examples

julia> D, (x, y), (dx, dy) = weyl_algebra(["x", "y"])
(2-d Weyl algebra in [x,y], PfaffianSystems.WAlgElem{AbstractAlgebra.Generic.MPoly{AbstractAlgebra.Generic.MPoly{Rational{BigInt}}}}[x, y], PfaffianSystems.WAlgElem{AbstractAlgebra.Generic.MPoly{AbstractAlgebra.Generic.MPoly{Rational{BigInt}}}}[dx, dy])
julia> evaluate(x*dx+y, [x], [x + y])
(x + y)*dx + y
julia> evaluate(x*dx + y, [dx], [dx+dy])
x*dx + x*dy + y

AbstractAlgebra.exponent_vectors — Method

exponent_vectors(f::T) where T <: AbstractDiffOp

Extract the exponent vectors of a differential operator.

Arguments:

f::T: A differential operator to extract the exponent vectors from.

Returns:

An array of arrays representing the exponent vectors of the differential operator.

AbstractAlgebra.monomials — Method

monomials(f::T) where T <: AbstractDiffOp

Extract the monomials of a differential operator.

Arguments:

f::T: A differential operator to extract the monomials from.

Returns:

An array of instances of T representing the monomials of the differential operator.

AbstractAlgebra.vars — Method

vars(wae::T) where T <: AbstractDiffOp

Examples

julia> D, (x, y), (dx, dy) = weyl_algebra(["x", "y"])
(2-d Weyl algebra in [x,y], PfaffianSystems.WAlgElem{AbstractAlgebra.Generic.MPoly{AbstractAlgebra.Generic.MPoly{Rational{BigInt}}}}[x, y], PfaffianSystems.WAlgElem{AbstractAlgebra.Generic.MPoly{AbstractAlgebra.Generic.MPoly{Rational{BigInt}}}}[dx, dy])
julia> vars(x+y)
2-element Vector{PfaffianSystems.WAlgElem{AbstractAlgebra.Generic.MPoly{AbstractAlgebra.Generic.MPoly{Rational{BigInt}}}}}:
 x
 y

PfaffianSystems.diff_op_ring — Method

Ring of differential operators over rational functions

PfaffianSystems.dvars — Method

dvars(wae::T) where T <: AbstractDiffOp

Examples

julia> D, (x, y), (dx, dy) = weyl_algebra(["x", "y"])
(2-d Weyl algebra in [x,y], PfaffianSystems.WAlgElem{AbstractAlgebra.Generic.MPoly{AbstractAlgebra.Generic.MPoly{Rational{BigInt}}}}[x, y], PfaffianSystems.WAlgElem{AbstractAlgebra.Generic.MPoly{AbstractAlgebra.Generic.MPoly{Rational{BigInt}}}}[dx, dy])
julia> dvars(dx^2+y)
1-element Vector{PfaffianSystems.WAlgElem{AbstractAlgebra.Generic.MPoly{AbstractAlgebra.Generic.MPoly{Rational{BigInt}}}}}:
 dx

PfaffianSystems.integration_DIdeal — Method

integration_DIdeal(I::DIdeal{T}, integ_vars::OrderedSet{T}) where T <: WAlgElem
integration_DIdeal(I::DIdeal{T}, integ_vars::Vector{T}) where T <: WAlgElem

Return the integration ideal of I with respect to integ_vars.

Examples

julia> D, (x, y), (dx, dy) = weyl_algebra(["x", "y"])
(2-d Weyl algebra in [x,y], PfaffianSystems.WAlgElem{AbstractAlgebra.Generic.MPoly{AbstractAlgebra.Generic.MPoly{Rational{BigInt}}}}[x, y], PfaffianSystems.WAlgElem{AbstractAlgebra.Generic.MPoly{AbstractAlgebra.Generic.MPoly{Rational{BigInt}}}}[dx, dy])
julia> I = DIdeal(D, [dx+x, dy+y])
Ideal of 2-d Weyl algebra in [x,y] generated by [dx + x, dy + y]
julia> integration_DIdeal(I, [x])
Ideal of 2-d Weyl algebra in [x,y] generated by [dy + y]

PfaffianSystems.intersection_DIdeals — Method

intersection_DIdeals(Is::Ideal...)

Return the intersection of the D-ideals Is[1], Is[2], ...

Examples

julia> D, x, dx = weyl_algebra("x")
(1-d Weyl algebra in [x], x, dx)
julia> I1 = DIdeal(D, [dx^2 + 1])
Ideal of 1-d Weyl algebra in [x] generated by [dx^2 + 1]
julia> I2 = DIdeal(D, [dx + x])
Ideal of 1-d Weyl algebra in [x] generated by [dx + x]
julia> intersection_DIdeals(I1, I2)
Ideal of 1-d Weyl algebra in [x] generated by [x*dx^3 + (x^2 - 2)*dx^2 + x*dx + x^2 - 2, dx^5 + x*dx^4 + 6*dx^3 + 2*x*dx^2 + 5*dx + x]

PfaffianSystems.isAsirAvailable — Method

Return the version of available Asir as an integer if exists, and otherwise return nothing.

PfaffianSystems.leading_term — Method

leading_term(f::DORElem, order::Symbol=:lex)

Return the leading term of f with respect to order. Only order=:lex and order=:grevlex are supported now. Default value of order is :grevlex.

PfaffianSystems.multiplication_DIdeal — Method

multiplication_DIdeal(I::DIdeal{T}, J::DIdeal{T}) where T <: WAlgElem

Return an ideal that annihilates the product fg where f and g are annihilated by I and J, respectively.

Exmaples

julia> D, x, dx = weyl_algebra("x")
(1-d Weyl algebra in [x], x, dx)
julia> I1 = DIdeal(D, [dx^2 + 1])
Ideal of 1-d Weyl algebra in [x] generated by [dx^2 + 1]
julia> I2 = DIdeal(D, [dx + x])
Ideal of 1-d Weyl algebra in [x] generated by [dx + x]
julia> multiplication_DIdeal(I1, I2)
Ideal of 1-d Weyl algebra in [x] generated by [-dx^2 - 2*x*dx - x^2 - 2]

PfaffianSystems.normalform — Method

normalform(f::T, G::Vector{T}) where T<: DORElem

Compute normal form of f with respect to G and return the remainder r and quotients q.

PfaffianSystems.parseAsir — Method

parseAsir(asir_res::String)

Add square brackets and separete the response into lines.

PfaffianSystems.pfaffian_system — Method

pfaffian_system(G::Vector{T}, S::Vector{T}) where T <: DORElem

Examples

julia> R, (x, y), (dx, dy) = diff_op_ring(["x", "y"])
(2-dimensional ring of differential opeartors in [x,y], PfaffianSystems.DORElem{AbstractAlgebra.Generic.MPoly{AbstractAlgebra.Generic.RationalFunctionFieldElem{Rational{BigInt}, AbstractAlgebra.Generic.MPoly{Rational{BigInt}}}}}[x, y], PfaffianSystems.DORElem{AbstractAlgebra.Generic.MPoly{AbstractAlgebra.Generic.RationalFunctionFieldElem{Rational{BigInt}, AbstractAlgebra.Generic.MPoly{Rational{BigInt}}}}}[dx, dy])
julia> pfaffian_system([dx^2 + 1, dy^2 + 1], [one(dx), dx, dy])
2-element Vector{Matrix{AbstractAlgebra.Generic.RationalFunctionFieldElem{Rational{BigInt}, AbstractAlgebra.Generic.MPoly{Rational{BigInt}}}}}:
 [0 1 0; -1 0 0; 0 0 0]
 [0 0 1; 0 0 0; -1 0 0]

PfaffianSystems.restriction_DIdeal — Method

restriction_DIdeal(I::DIdeal{T}, rest_vars::OrderedSet{T}) where T <: WAlgElem
restriction_DIdeal(I::DIdeal{T}, rest_vars::Vector{T}) where T <: WAlgElem

Return the restriction ideal of I with respect to the subspace where every variable of rest_vars equals to zero.

Examples

julia> D, (x, y), (dx, dy) = weyl_algebra(["x", "y"])
(2-d Weyl algebra in [x,y], PfaffianSystems.WAlgElem{AbstractAlgebra.Generic.MPoly{AbstractAlgebra.Generic.MPoly{Rational{BigInt}}}}[x, y], PfaffianSystems.WAlgElem{AbstractAlgebra.Generic.MPoly{AbstractAlgebra.Generic.MPoly{Rational{BigInt}}}}[dx, dy])
julia> I = DIdeal(D, [dx^2 + 2*dx*dy + dy^2 + 1, -dy + x - y])
Ideal of 2-d Weyl algebra in [x,y] generated by [dx^2 + 2*dx*dy + dy^2 + 1, -dy + x - y]
julia> restriction_DIdeal(I, [y])
Ideal of 2-d Weyl algebra in [x,y] generated by [-dx^2 - 2*x*dx - x^2 - 2]

PfaffianSystems.runAsir — Method

runAsir(commands::AbstractString)

Run commands on Asir. The raw response of Asir is returned as a string.

PfaffianSystems.stdmon — Method

stdmon!(I::DIdeal[, ordered_vars::Vector{Num}])

Compute standard monomials if I is zero-dimensional as an ideal of R-ring. The order of varaibles in the computation of Grobner bases can be provided as ordered_vars, which can be omitted as long as the computational time is not crutial.

PfaffianSystems.vec2str — Method

vec2str(v::AbstractVector; delim=",")
vec2str(v::AbstractDiffOp; delim=",")
vec2str(vs...; delim=",")

Return a string consiting of all elements of v with delimiter delim.

PfaffianSystems.weyl_algebra — Method

WeylAlgebra

PfaffianSystems.wnormalform — Method

wnormalform(f::T, G::Vector{T}) where T<: DORElem Compute weak normal form of f with respect to G and return the remainder r and quotients q.