PfaffianSystems

Documentation for PfaffianSystems.

PfaffianSystems.DIdeal

PfaffianSystems.PfaffianSystem

PfaffianSystems.addVars

PfaffianSystems.apply_doterm

PfaffianSystems.eliminationIdeal

PfaffianSystems.genVars

PfaffianSystems.intersectionIdeal

PfaffianSystems.isAsirAvailable

PfaffianSystems.isZeroDimensional

PfaffianSystems.parseAsir

PfaffianSystems.runAsir

PfaffianSystems.stdmon!

PfaffianSystems.vec2str

PfaffianSystems.DIdeal — Type

DIdeal <: AbstractIdeal

D-Ideal type

PfaffianSystems.PfaffianSystem — Type

Pfaffian system

PfaffianSystems.addVars — Method

addVars(name::AbstractString[, n::Integer, v2d::Bijection])

Generate a Symbolics variable name and its corresponding differential operator dname. When n is given, generate vectors [name1...namen] and [dname1...dnamen]. The correspondence between name and dname is added to the bijection v2d.

Examples

julia> x, dx, v2d = genVars("x")
(x, dx, Bijection{Symbolics.Num,Symbolics.Num} (with 1 pairs))

julia> y, dy, v2d = addVars("y", 2, v2d)
(Symbolics.Num[y1, y2], Symbolics.Num[dy1, dy2], Bijection{Symbolics.Num,Symbolics.Num} (with 3 pairs))

PfaffianSystems.apply_doterm — Method

apply_dmon(DOmon::AbstractTerm, F::Num, p2s::Bijection, v2d::Bijection)

Apply a term of differential operator DOterm to an expression F. The bijection p2s relates a MulltivariatePolunomials expression to a SymbolicUtils one, and d2v does a differential operator to its corresponding variable.

PfaffianSystems.eliminationIdeal — Method

eliminationIdeal(I::DIdeal, elim_vars::OrderedSet{Num})
eliminationIdeal(I::DIdeal, elim_vars::Vector{Num})

Return elimination ideal of I that does not include any variable of elim_vars.

PfaffianSystems.genVars — Method

genVars(name::AbstractString[, n::Integer])

Generate a Symbolics variable name and its corresponding differential operator dname. When n is given, generate vectors [name1...namen] and [dname1...dnamen]. The correspondence between name and dname is returned as a new bijection v2d, which will be provided to addVars.

Examples

julia> x, dx, v2d = genVars("x")
(x, dx, Bijection{Symbolics.Num,Symbolics.Num} (with 1 pairs))

julia> y, dy, v2d = addVars("y", 2, v2d)
(Symbolics.Num[y1, y2], Symbolics.Num[dy1, dy2], Bijection{Symbolics.Num,Symbolics.Num} (with 3 pairs))

PfaffianSystems.intersectionIdeal — Method

intersectionIdeal(Is::DIdeal...)

Return the intersection of the D-ideals Is[1], Is[2], ...

PfaffianSystems.isAsirAvailable — Method

Return the version of available Asir as an integer if exists, and otherwise return nothing.

PfaffianSystems.isZeroDimensional — Method

isZeroDimensional(I::DIdeal)

Check if the ideal in R generated by DIdeal I is zero-dimensional.

PfaffianSystems.parseAsir — Method

parseAsir(asir_res::String)

Add square brackets and separete the response into lines.

PfaffianSystems.runAsir — Method

runAsir(commands::AbstractString)

Run commands on Asir. The raw response of Asir is returned as a string.

PfaffianSystems.stdmon! — Method

stdmon!(I::DIdeal[, ordered_vars::Vector{Num}])

Compute standard monomials if I is zero-dimensional as an ideal of R-ring. The order of varaibles in the computation of Grobner bases can be provided as ordered_vars, which can be omitted as long as the computational time is not crutial.

PfaffianSystems.vec2str — Method

vec2str(v::Vector{Num}; delim=",")
vec2str(v::Symbolics.Arr; delim=",")
vec2str(v::OrderedSet{Num}; delim=",")

Return a string consiting of all elements of v with delimiter delim.